+++ MatrixKit/Models/Room/MXKRoomBubbleCellData.h
@@ -1,5 +1,6 @@
 /*
  Copyright 2015 OpenMarket Ltd
+ Copyright 2019 The Matrix.org Foundation C.I.C
 
  Licensed under the Apache License, Version 2.0 (the "License");
  you may not use this file except in compliance with the License.
@@ -68,6 +69,11 @@
  */
 @property(nonatomic) NSMutableDictionary<NSString* /* eventId */,
                                          NSArray<MXReceiptData *> *> *readReceipts;
+
+/**
+ Aggregated reactions per event.
+ */
+@property(nonatomic) NSMutableDictionary<NSString* /* eventId */, MXAggregatedReactions*> *reactions;
 
 /**
  Event formatter